pyproject.toml
[tool.poetry]
name = "cookie-stand-api"
version = "0.1.0"
description = ""
authors = ["Yahialabeeb <yahia.labeeb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
django-environ = "^0.8.1"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
